+++ skoolkit/defaults.py
@@ -124,6 +124,7 @@ SECTIONS['MemoryMap:MemoryMap'] = """
 EntryDescriptions=0
 EntryTypes=bcgstuw
 Intro=
+LengthColumn=0
 PageByteColumns=1
 Write=1
 """
@@ -132,6 +133,7 @@ SECTIONS['MemoryMap:RoutinesMap'] = """
 EntryDescriptions=0
 EntryTypes=c
 Intro=
+LengthColumn=0
 PageByteColumns=0
 Write=1
 """
@@ -140,6 +142,7 @@ SECTIONS['MemoryMap:DataMap'] = """
 EntryDescriptions=0
 EntryTypes=bw
 Intro=
+LengthColumn=0
 PageByteColumns=1
 Write=1
 """
@@ -148,6 +151,7 @@ SECTIONS['MemoryMap:MessagesMap'] = """
 EntryDescriptions=0
 EntryTypes=t
 Intro=
+LengthColumn=0
 PageByteColumns=0
 Write=1
 """
@@ -156,6 +160,7 @@ SECTIONS['MemoryMap:UnusedMap'] = """
 EntryDescriptions=0
 EntryTypes=suz
 Intro=
+LengthColumn=0
 PageByteColumns=1
 Write=1
 """